requirements.txt
alembic==1.3.2
aniso8601==6.0.0
astroid==2.2.5
awscli==1.17.15
Babel==2.8.0
botocore==1.14.15
Click==7.0
colorama==0.4.3
docutils==0.15.2
ecdsa==0.13.3
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
future==0.17.1
gunicorn==20.0.4
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
lazy-object-proxy==1.4.0
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
pyasn1==0.4.8
pycryptodome==3.6.6
pylint==2.3.1
python-dateutil==2.6.0
python-editor==1.0.4
python-jose-cryptodome==1.3.2
pytz==2019.1
PyYAML==5.2
rsa==3.4.2
s3transfer==0.3.3
six==1.12.0
SQLAlchemy==1.3.13
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.1
WTForms==2.2.1
